fix: Client-side tile deletion out of order
Previously we were sending 3 packets to delete the three tiles. They were sent and arrived in an arbitrary order.
Locally to the GM all looks fine, but client side there is a flicker effect as they are deleted in an arbitrary order with packet delay.

Update to only use one packet to instruct the server to delete all the tiles. The clients should then pull the update from the server and also delete all the tiles simeoutaneously.

+++ src/animation/effects/token-mask/token-mask.js
@@ -144,15 +144,8 @@ async function create(token, config = {}) {
     seq = seq.waitUntilFinished()
 
         .thenDo(async () => {
-            if (deleteToken) {
-                await token.document.delete();
-            } else {
-                await Promise.all([
-                    socket.tile.destroy(tokenRevealMask.id),
-                    socket.tile.destroy(tokenShapeMask.id),
-                    socket.tile.destroy(sceneRevealMask.id),
-                ]);
-            }
+            await socket.tile.destroy([tokenRevealMask.id, tokenShapeMask.id, sceneRevealMask.id]);
+            if (deleteToken) await token.document.delete();
             await Sequencer.EffectManager.endEffects({ name: label });
         });
